(PA-1824)(PA-2141) Update ruby builds to dynamically generate rbconfigs
Previously, ruby builds would copy in static files to replace the rbconfigs
generated during the ruby build. This was very error prone and pretty annoying
to do every time we updated ruby.

This commit replaces that process with a ruby script that uses either the
newly built ruby or pl-ruby (in XCC situations) to regenerate the rbconfig
with any changes requested.

The new ruby script, rbconfig-update.rb is copied to the builder and then
after the ruby build is run. It will generate two files: a copy of the
original rbconfig with exceptions changed to warnings (this is used by
downstream components to build things, and in cross compiled situations will
fail the build because one of the exceptions will raise) and a new rbconfig:
new_rbconfig.rb which is a version of the original rbconfig with any
replacements listed by the component.

new_rbconfig once generated replaces the rbconfig in the actual ruby we build.